requirements.txt
```
pandas
numpy
pyarrow
scikit-learn
sentence-transformers
faiss-cpu
gradio
transformers
```